Fix bf16 models (load float32) + representative reference set
- Load models in float32: Qwen (and others) default to bfloat16, which has no numpy
  bridge — manifold/embeddings crashed ("unsupported ScalarType BFloat16"). float32 is
  also exact and no slower on CPU. Regression test asserts the float32 invariant.
- Reference token set is now evenly spaced across the vocabulary instead of the first N
  ids (which are mostly byte/special tokens), so the manifold/vector field cover a
  representative, interesting spread of tokens.
- Bump cache SCHEMA_VERSION (1->2) to invalidate artifacts computed under the old dtype
  / reference set. 50 backend tests pass.

+++ code/backend/src/llm_geometry/compute/manifold.py
@@ -77,7 +77,7 @@ def manifold(
 
     lm = load_model(model_id)
     token_ids = reference_token_ids(lm, reference_set_size)
-    matrix = lm.model.get_input_embeddings().weight.detach().cpu().numpy().astype(np.float64)
+    matrix = lm.model.get_input_embeddings().weight.detach().float().cpu().numpy().astype(np.float64)
     emb = matrix[token_ids]
 
     from ..reduce.sphere import reduce_3d_sphere
